Implement purchase-related info window and Mediator pattern
Implemented the functionality of displaying purchase-related information window when purchasing a product. Also, implemented the Mediator pattern to update the quantity of cooking ingredients in real-time when they are added, and to manage the gain and use of in-game currency. Singleton classes were assigned to each object to clarify their roles and unnecessary code was removed.